requirements.txt
# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# ML
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.0
joblib==1.3.2


# Airflow providers
apache-airflow-providers-postgres==5.10.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
geopy

# Test
pytest